// moonbowcl/src/app/model/dto/form-iomodulo.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

export class FormIOModulo {
  // tslint:disable-next-line:variable-name
  _id: string;
  type: string;
  name: string;
  title: string;
  path: string;
  display: string;
  components: [];
  machineName: string;
  // tslint:disable-next-line:variable-name
  __v: string;
}